+++ conman/redirects/handlers.py
@@ -1,7 +1,0 @@
-from conman.routes.handlers import SimpleHandler
-from . import views
-
-
-class RouteRedirectHandler(SimpleHandler):
-    """Pass a request through to RouteRedirectView."""
-    view = views.RouteRedirectView.as_view()